+++ assets/js/app/core/auth/services/UserService.js
@@ -22,43 +22,60 @@
  *
  * Happy coding!
  */
-(function() {
+(function () {
   'use strict';
 
   angular.module('frontend.core.auth.services')
     .factory('UserService', [
-        '_','$localStorage','$rootScope','AuthService',
-      function factory(_,$localStorage,$rootScope, AuthService) {
+      '_', '$localStorage', '$rootScope', 'AuthService',
+      function factory(_, $localStorage, $rootScope, AuthService) {
 
-          function user() {
+        function user() {
+          var user = $localStorage.credentials ? $localStorage.credentials.user : {};
+          if(window.no_auth) {
+            var guestUser = {
+              id: 999,
+              username: 'guest',
+              email: 'guest@guest.com',
+              firstName: 'Guest',
+              lastName: 'Guest',
+              admin: true,
+              active: true
+            };
 
-              var user = $localStorage.credentials ? $localStorage.credentials.user : {};
-
-              if(user.id) {
-                  user.hasPermission =  AuthService.hasPermission
-              }
-
-
-              return user;
+            if(!user.id || user.id !== guestUser.id || user.username !== guestUser.username || user.email !== guestUser.email) {
+              $localStorage.credentials = {
+                user: guestUser,
+                token: 'noauthtoken'
+              };
+            }
           }
 
-          function updateUser(user,keepNode) {
-
-              var existingUser = $localStorage.credentials ? $localStorage.credentials.user : {};
-
-              // Only update localStorage if user is the same
-              if(existingUser.id == user.id) {
-                  if(keepNode) {
-                      user.node =$localStorage.credentials.user.node // Retain user node
-                  }
-                  $localStorage.credentials.user = user;
-                  $rootScope.$broadcast('user.updated',$localStorage.credentials.user)
-              }
+          if (user.id) {
+            user.hasPermission = AuthService.hasPermission
           }
+
+
+          return user;
+        }
+
+        function updateUser(user, keepNode) {
+
+          var existingUser = $localStorage.credentials ? $localStorage.credentials.user : {};
+
+          // Only update localStorage if user is the same
+          if (existingUser.id == user.id) {
+            if (keepNode) {
+              user.node = $localStorage.credentials.user.node // Retain user node
+            }
+            $localStorage.credentials.user = user;
+            $rootScope.$broadcast('user.updated', $localStorage.credentials.user)
+          }
+        }
 
         return {
-            user: user,
-            updateUser : updateUser
+          user: user,
+          updateUser: updateUser
         };
       }
     ])