+++ assets/js/app/healthchecks/healthchecks-controller.js
@@ -192,7 +192,7 @@
         $scope.deleteItem = function deleteItem(item) {
 
 
-          DialogService.prompt(
+          DialogService.confirm(
             "Confirm", "Really want to delete the selected item?",
             ['No don\'t', 'Yes! delete it'],
             function accept() {