+++ assets/js/app/plugins/add-plugin-modal-controller.js
@@ -30,7 +30,7 @@
 
         // Define the plugins that will have their own custom form
         // so that it can be included via ng-include in the .html files
-        $scope.customPluginForms = ['statsd'];
+        $scope.customPluginForms = ['statsd', 'response-ratelimiting'];
 
         new KongPluginsService().makePluginGroups().then(function (groups) {
           $scope.pluginGroups = groups