+++ assets/js/app/routes/controllers/add-route-modal-controller.js
@@ -30,6 +30,14 @@
           $uibModalInstance.dismiss()
         }
 
+        $scope.onTagInputKeyPress = function ($event) {
+          if($event.keyCode === 13) {
+            if(!$scope.route.tags) $scope.route.tags = [];
+            $scope.route.tags = $scope.route.tags.concat($event.currentTarget.value);
+            $event.currentTarget.value = null;
+          }
+        }
+
 
         $scope.submit = function () {